+++ database/migrations/2025_06_07_000003_create_companies_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
             $table->text('description')->nullable();
             $table->string('slug');
             $table->string('country')->nullable();
-            $table->integer('wage_strike')->default(1);
+            $table->float('wage_strike')->default(1);
             $table->timestamps();
         });
     }
